requirements.txt
flask==3.0.2
yt-dlp==2023.12.30
replicate==0.22.0
python-dotenv==1.0.1
gunicorn==21.2.0